refactor: organize Ads components
