+++ components/Ads/index.ts
@@ -1,0 +1,3 @@
+export { default as Placeholder } from "./Placeholder";
+export { default as TheMoneytizer } from "./TheMoneytizer";
+export { default as Vidverto } from "./Vidverto";
