+++ components/CategoryData/CategoryData.tsx
@@ -1,6 +1,6 @@
 import { ICategoryData } from "@mauriciorobayo/pyptron";
 import Category from "@mauriciorobayo/pyptron/dist/models/category";
-import Vidverto from "components/ads/vidverto";
+import { Vidverto } from "components/Ads";
 import { memo, useEffect, useState } from "react";
 import NumberLinks from "../number-links";
 import {
